Move the embedded error-mapping test to the suite that has an engine
CI failed on all 24 Core Tests legs - every server version and every Python -
while both of my local runs passed. The test asserting that a closed embedded
connection raises ConnectionUnavailableError uses Surreal("mem://"), which
needs the optional native engine. My local venv has it because I build it with
maturin; the core CI job installs no engine at all.

That is the same blind spot this run of work has been about: a test assuming
the embedded engine is present, exactly as __all__ used to.

A skip guard would not fix it. The core job would skip the test, and the
embedded job runs only tests/unit_tests/connections/embedded, so a guarded
test in the shared file would never execute in CI anywhere. Move it to the
embedded directory instead, where the job that builds the engine will run it
across Python 3.10-3.13.

Verified under the failing condition this time rather than assuming: an
engine-free venv running the core job's exact command
(--ignore=tests/unit_tests/connections/embedded) gives 1035 passed, the
embedded job's command gives 28 passed, and the full local suite with an
engine gives 1063.
